Move project status types enum in separate file
