+++ src/app/dataModels/project-status-types.ts
@@ -1,0 +1,6 @@
+export enum ProjectStatusTypes {
+  active = 'Active',
+  completed = 'Completed',
+  onHold = 'On Hold',
+  inactive = 'Inactive',
+}
